Restructure session and token function + adding last middleware tests
If a function is both used externally and internally in a module it can
be difficult to mock everything. The solution is to separate the
functions into the own "domains". Session handling in one module,
library tokens in its own module, user tokens in its own module and so
forth.
