+++ app/auth/session/route.ts
@@ -7,12 +7,7 @@ export async function GET() {
     if (!session) {
       return Response.json({ defaultSession })
     }
-    return Response.json({
-      isLoggedIn: session.isLoggedIn,
-      userInfo: session.userInfo,
-      type: session.type,
-      adgangsplatformenUserToken: session.adgangsplatformenUserToken,
-    })
+    return Response.json(session)
   } catch (e) {
     return Response.json({ error: e }, { status: 500 })
   }
